Jan 20 18:34pm Qin

+++ src/model/TransactionDAO.java
@@ -40,7 +40,7 @@ public class TransactionDAO extends BaseDAO{
 		return true;
 	}
 	
-	public void buyFund(int customerId, int fundId, long amount) {
+	public void buyFund(int customerId, int fundId, double amount) {
 		// some data will be stored in the transaction table
 		
 		// the balance of customer will be updated
